Updated components
Now it's possible to obtain servers TPS due the API has been fixed, now exceptions are handled by the user, removed the usage of UUID utils class.

+++ src/main/java/com/github/grubsic/rcapi/interfaces/Server.java
@@ -9,7 +9,7 @@ public interface Server{
 	boolean isOnline();
 	int getRAMUsage();
 	int getAvailableRAM();
-	int getTps();
+	double getTps();
 	String getServerGameVersion();
 	boolean isStaffServer();
 	int getCurrentPlayers();
